Renamed Player.id to playerId/added packages for sorting
