+++ src/main/kotlin/com/aau/se2/boomboomkittens/filipp/server/controllers/rest/LobbyRestController.kt
@@ -1,7 +1,7 @@
-package com.aau.se2.boomboomkittens.filipp.server.controllers.REST
+package com.aau.se2.boomboomkittens.filipp.server.controllers.rest
 
 import com.aau.se2.boomboomkittens.filipp.server.models.Lobby
-import com.aau.se2.boomboomkittens.filipp.server.models.Player
+import com.aau.se2.boomboomkittens.filipp.server.models.player.Player
 import com.aau.se2.boomboomkittens.filipp.server.services.LobbyService
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
