+++ src/main/kotlin/com/aau/se2/boomboomkittens/filipp/server/models/Lobby.kt
@@ -1,5 +1,6 @@
 package com.aau.se2.boomboomkittens.filipp.server.models
 
+import com.aau.se2.boomboomkittens.filipp.server.models.player.Player
 import java.util.UUID
 
 data class Lobby(val id:UUID = UUID.randomUUID(),
