--- src/main/kotlin/com/aau/se2/boomboomkittens/filipp/server/models/Player.kt
@@ -1,5 +1,0 @@
-package com.aau.se2.boomboomkittens.filipp.server.models
-
-import java.util.UUID
-
-data class Player(val id: UUID? = UUID.randomUUID(), val name:String)
