+++ src/test/kotlin/com/aau/se2/boomboomkittens/filipp/server/controllers/rest/LobbyRestControllerTest.kt
@@ -1,7 +1,7 @@
-package com.aau.se2.boomboomkittens.filipp.server.controllers.REST
+package com.aau.se2.boomboomkittens.filipp.server.controllers.rest
 
 import com.aau.se2.boomboomkittens.filipp.server.models.Lobby
-import com.aau.se2.boomboomkittens.filipp.server.models.Player
+import com.aau.se2.boomboomkittens.filipp.server.models.player.Player
 import com.aau.se2.boomboomkittens.filipp.server.services.LobbyService
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
