+++ src/test/kotlin/com/aau/se2/boomboomkittens/filipp/server/controllers/rest/PlayerRestControllerTest.kt
@@ -1,6 +1,6 @@
-package com.aau.se2.boomboomkittens.filipp.server.controllers.REST
+package com.aau.se2.boomboomkittens.filipp.server.controllers.rest
 
-import com.aau.se2.boomboomkittens.filipp.server.models.Player
+import com.aau.se2.boomboomkittens.filipp.server.models.player.Player
 import com.aau.se2.boomboomkittens.filipp.server.services.PlayerService
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.junit.jupiter.api.Test
@@ -37,20 +37,20 @@ class PlayerRestControllerTest {
         val player = Player(id,"Dummy")
 
 
-        given(playerService.getPlayer(player.id.toString())).willReturn(player)
+        given(playerService.getPlayer(player.playerId.toString())).willReturn(player)
 
         mockMvc.get("/players"){
             header("id",id)
         }.andExpect {
             status { isOk() }
             jsonPath("$.name"){value("Dummy")}
-            jsonPath("$.id"){value(id.toString())}
+            jsonPath("$.playerId"){value(id.toString())}
         }
         }
 
     @Test
     fun registerPlayerTest(){
-        val player = Player(null, "Dummy")
+        val player = Player(UUID.randomUUID(), "Dummy")
         val playerJson = objectMapper.writeValueAsString(player)
