+++ src/test/kotlin/com/aau/se2/boomboomkittens/filipp/server/services/LobbyServiceTest.kt
@@ -1,6 +1,6 @@
 package com.aau.se2.boomboomkittens.filipp.server.services
 
-import com.aau.se2.boomboomkittens.filipp.server.models.Player
+import com.aau.se2.boomboomkittens.filipp.server.models.player.Player
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
